+++ packages/commercetools/api-client/src/index.ts
@@ -24,9 +24,6 @@ import customerUpdateMe from './api/customerUpdateMe';
 import createAccessToken from './helpers/createAccessToken';
 import { apiClientFactory } from '@vue-storefront/core';
 import { Config, ConfigurableConfig } from './types/setup';
-export * from './types/Api';
-export * from './types/setup';
-export * as cartActions from './helpers/cart/actions';
 
 let apolloClient: ApolloClient<any> = null;
 
@@ -80,3 +77,9 @@ export {
   customerChangeMyPassword,
   customerUpdateMe
 };
+
+export * from './types/Api';
+export * from './types/GraphQL';
+export * from './types/setup';
+export * from './helpers/token';
+export * as cartActions from './helpers/cart/actions';