+++ packages/core/core/src/factories/useProductFactory.ts
@@ -1,6 +1,6 @@
 import { UseProduct } from '../types';
 import { Ref, computed } from '@vue/composition-api';
-import { sharedRef } from '../utils';
+import { sharedRef, Logger } from '../utils';
 import { CustomQuery } from '../types';
 
 type SearchParams = {
@@ -33,6 +33,8 @@ export function useProductFactory<PRODUCT, PRODUCT_SEARCH_PARAMS, PRODUCT_FILTER
     const loading = sharedRef(false, `useProduct-loading-${id}`);
 
     const search = async (params: PRODUCT_SEARCH_PARAMS, customQuery?: CustomQuery) => {
+      Logger.debug('useProduct.search', params);
+
       loading.value = true;
       filters.value = null;
       try {
@@ -41,6 +43,8 @@ export function useProductFactory<PRODUCT, PRODUCT_SEARCH_PARAMS, PRODUCT_FILTER
         totalProducts.value = total;
         filters.value = availableFilters || null;
         sortingOptions.value = availableSortingOptions || null;
+      } catch (e) {
+        Logger.error('useProduct.search', e);
       } finally {
         loading.value = false;
       }