+++ packages/core/docs/commercetools/changelog.md
@@ -2,11 +2,12 @@
 
 ## 0.2.0 (not released)
 
+- Fix types for CT api-client and composables packages ([#4924](https://github.com/DivanteLtd/vue-storefront/pull/4924))
 - fixed updateUser on useUser composable ([#4863](https://github.com/DivanteLtd/vue-storefront/issues/4863))
 - implemented useReviews on product page ([#4800](https://github.com/DivanteLtd/vue-storefront/issues/4800))
 - implemented faceting using useFacet factory (#4853)
+- fixed anonymous token loading (#4917)
 
 ## 0.1.0
 
 - refactored setup using apiClientFactory ([#4777](https://github.com/DivanteLtd/vue-storefront/issues/4777))
-